move patchify to outline transforms
