+++ tests/transforms/outline/test_patchify.py
