+++ torchfont/transforms/__init__.py
@@ -19,8 +19,7 @@ from torchfont.transforms.geometric import (
     random_vertical_flip,
     vertical_flip,
 )
-from torchfont.transforms.outline import remove_overlaps
-from torchfont.transforms.patch import patchify
+from torchfont.transforms.outline import patchify, remove_overlaps
 
 __all__ = [
     "BitmapMode",
